Add REST API for adding colors
