+++ python_anywhere/bouldern/serializers.py
@@ -1,7 +1,7 @@
 """Serializers for the bouldern app"""
 from rest_framework.serializers import ModelSerializer
 
-from python_anywhere.bouldern.models import Gym
+from python_anywhere.bouldern.models import Gym, Color
 
 
 class UGCSerializer(ModelSerializer):
@@ -17,3 +17,10 @@ class GymSerializer(UGCSerializer):
     class Meta:
         model = Gym
         fields = ['name', 'map']
+
+
+class ColorSerializer(UGCSerializer):
+    """Serializer for Color instances"""
+    class Meta:
+        model = Color
+        fields = ['name', 'color']
