+++ python_anywhere/bouldern/urls.py
@@ -2,7 +2,8 @@
 from django.urls import path
 from django.views.generic import TemplateView
 
-from python_anywhere.bouldern.views import gym_map, AddGym, AddColor, AddGymRest
+from python_anywhere.bouldern.views import gym_map, AddGym, AddColor, \
+    AddGymRest, AddColorRest
 
 urlpatterns = [
     path('', TemplateView.as_view(template_name='bouldern/index.html'),
@@ -11,4 +12,5 @@ urlpatterns = [
     path('add-color/', AddColor.as_view(), name=AddColor.name),
     path('add-gym/', AddGym.as_view(), name=AddGym.name),
     path('rest/add-gym/', AddGymRest.as_view(), name=AddGymRest.name),
+    path('rest/add-color/', AddColorRest.as_view(), name=AddColorRest.name),
 ]
